Category refactor: code cleanup, moved CategoryList components to subfolders
