+++ webapp/views/App/views/Designer/CategoryList/TableHeaderLeft/index.js
@@ -1,0 +1,1 @@
+export { default } from './TableHeaderLeft'
