+++ webapp/views/App/views/Designer/CategoryList/TableRowHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from './TableRowHeader'
